dernieres modifs

+++ CPP06/ex01/Serializer.cpp
@@ -6,7 +6,7 @@
 /*   By: juandrie <juandrie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/23 17:10:42 by julietteand       #+#    #+#             */
-/*   Updated: 2024/03/25 11:27:01 by juandrie         ###   ########.fr       */
+/*   Updated: 2024/06/24 15:14:53 by juandrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,12 +31,12 @@ Serializer::~Serializer()
 {
 }
 
-unsigned long Serializer::serialize(Data *ptr)
+uintptr_t Serializer::serialize(Data *ptr)
 {
-	return (reinterpret_cast<unsigned long>(ptr));
+	return (reinterpret_cast<uintptr_t>(ptr));
 }
 
-Data *Serializer::deserialize(unsigned long raw)
+Data *Serializer::deserialize(uintptr_t raw)
 {
 	return (reinterpret_cast<Data*>(raw));
 }
